requirements.txt
dnspython==2.7.0
colorama==0.4.6
python-dotenv==1.0.1
langchain==0.2.16
langchain-groq==0.1.10
langchain-community==0.2.13
langchain-huggingface==0.0.3
faiss-cpu==1.9.0
fpdf2==2.8.1
tqdm==4.67.1